requirements.txt
fastapi
gunicorn
rembg
pillow
onnxruntime
python-multipart
uvicorn
